refactor: Reorganize ratelimitbypass package to filters/ratelimit/bypass
- Move bypass.go and bypass_test.go from ratelimitbypass/ to filters/ratelimit/bypass/
- Update package declaration from 'ratelimitbypass' to 'bypass'
- Update all import paths across the codebase:
  * filters/ratelimit/token.go
  * filters/ratelimit/token_test.go
  * filters/ratelimit/ratelimit.go
  * filters/ratelimit/ratelimit_test.go
  * skipper.go
- Update all type references (BypassValidator, BypassConfig, etc.)
- Update documentation in docs/tutorials/ratelimit.md
- Remove old ratelimitbypass/ directory
